Moved the auto code
