+++ src/main/java/frc/robot/commands/CMD_AutoDrive.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.Arcade;
+package frc.robot.commands;
 import frc.robot.subsystems.SUB_Drivetrain;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.Constants;
